+++ CHANGELOG.md
@@ -14,7 +14,9 @@
 ## 0.5.1
 
 - Apply breaking changes from 0.5.0 ([#100], [#130]) to the webclient ([#132])
+- Add API-endpoints for list of indicators, reports, layers, datasets and fid_field ([#106])
 
+[#106]: https://github.com/GIScience/ohsome-quality-analyst/issues/106
 [#132]: https://github.com/GIScience/ohsome-quality-analyst/pull/132